ui: test fixes
