+++ setup.py
@@ -73,6 +73,7 @@ install_requires = [
     "arrow>=0.15.0",
     "Flask-BabelEx>=0.9.3",
     "Flask-Debugtoolbar>=0.10.1",
+    "Flask-Login==0.4.1",
     "invenio[base,auth]==3.2.0a9",
     # `metadata` bundle without records UI
     "invenio-indexer>=1.1.0,<1.2.0",
@@ -98,6 +99,7 @@ install_requires = [
     # namedtuple are json serialized as dict
     "simplejson>=3",
     "webargs>=5.5.2,<6.0",
+    "Werkzeug==0.16.1",
 ]
 
 packages = find_packages()
