+++ ui/src/components/ESSelector/HitsSearch.js
@@ -59,6 +59,9 @@ export class HitsSearch extends Component {
         for (let hit of response.data.hits) {
           results.push(serialize(hit));
         }
+        if (this.props.onResults) {
+          this.props.onResults(results);
+        }
       } else {
         results = response.data.hits;
       }
