+++ ui/src/pages/backoffice/Document/DocumentDetails/DocumentDetails.js
@@ -1,172 +1,15 @@
-import { toShortDate } from '@api/date';
-import {
-  DocumentCover,
-  DocumentTags,
-  DocumentTitle,
-} from '@components/Document';
-import { DetailsHeader } from '@pages/backoffice';
-import { DocumentEItems } from './components';
-import {
-  DocumentSeries,
-  DocumentSiblings,
-} from './components/DocumentRelations';
+import { DocumentContent } from './DocumentContent';
+import { DocumentHeader } from './DocumentHeader';
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import {
-  Accordion,
-  Container,
-  Divider,
-  Grid,
-  Header,
-  Icon,
-  Label,
-  Ref,
-  Sticky,
-} from 'semantic-ui-react';
+import { Container, Divider, Grid, Ref, Sticky } from 'semantic-ui-react';
 import history from '@history';
-import { Loader, Error, CopyButton, DocumentAuthors } from '@components';
+import { Loader, Error } from '@components';
 import {
   DocumentMetadata,
-  DocumentPendingLoans,
-  DocumentStats,
-  DocumentItems,
   DocumentActionMenu,
   DocumentSummary,
-  DocumentCreatedBy,
 } from './components';
-
-class DocumentHeader extends Component {
-  render() {
-    const { data } = this.props;
-    const recordInfo = (
-      <>
-        <label className="muted">Document</label> {data.metadata.pid}{' '}
-        <CopyButton text={data.metadata.pid} />
-        {data.metadata.created_by && (
-          <>
-            <br />
-            <label className="muted">Created by</label>{' '}
-            <DocumentCreatedBy metadata={data.metadata} />
-          </>
-        )}
-        <br />
-        <label className="muted">Created on</label> {toShortDate(data.created)}
-        <br />
-        {!data.metadata.open_access && (
-          <Label size="large" icon color="red">
-            <Icon name="lock" />
-            Restricted access
-          </Label>
-        )}
-      </>
-    );
-    return (
-      <DetailsHeader
-        title={
-          <>
-            <Header.Subheader>{data.metadata.document_type}</Header.Subheader>
-            <DocumentTitle document={data} />
-          </>
-        }
-        subTitle={<DocumentAuthors metadata={data.metadata} prefix={'by '} />}
-        pid={data.metadata.pid}
-        icon={<DocumentCover document={data} imageSize="huge" />}
-        recordType="Document"
-        recordInfo={recordInfo}
-      >
-        <DocumentTags metadata={data.metadata} />
-      </DetailsHeader>
-    );
-  }
-}
-
-class DocumentContent extends Component {
-  render() {
-    const { anchors } = this.props;
-    const panels = [
-      {
-        key: 'loan-requests',
-        title: 'Loan requests',
-        content: (
-          <Accordion.Content>
-            <div ref={anchors.loanRequestsRef} id="loan-requests">
-              <DocumentPendingLoans />
-            </div>
-          </Accordion.Content>
-        ),
-      },
-      {
-        key: 'document-items',
-        title: 'Physical items',
-        content: (
-          <Accordion.Content>
-            <div ref={anchors.attachedItemsRef} id="document-items">
-              <DocumentItems />
-            </div>
-          </Accordion.Content>
-        ),
-      },
-      {
-        key: 'document-eitems',
-        title: 'Electronic items',
-        content: (
-          <Accordion.Content>
-            <div ref={anchors.attachedEItemsRef} id="document-eitems">
-              <DocumentEItems />
-            </div>
-          </Accordion.Content>
-        ),
-      },
-      {
-        key: 'document-series',
-        title: 'Series',
-        content: (
-          <Accordion.Content>
-            <div ref={anchors.seriesRef} id="document-series">
-              <DocumentSeries />
-            </div>
-          </Accordion.Content>
-        ),
-      },
-      {
-        key: 'document-siblings',
-        title: 'Related',
-        content: (
-          <Accordion.Content>
-            <div ref={anchors.relatedRef} id="document-siblings">
-              <DocumentSiblings />
-            </div>
-          </Accordion.Content>
-        ),
-      },
-      {
-        key: 'document-statistics',
-        title: 'Statistics',
-        content: (
-          <Accordion.Content>
-            <div ref={anchors.statisticsRef} id="document-statistics">
-              <DocumentStats />
-            </div>
-          </Accordion.Content>
-        ),
-      },
-    ];
-    const defaultIndexes = [0, 1, 3];
-
-    return (
-      <>
-        <Accordion
-          fluid
-          styled
-          className="highlighted"
-          panels={panels}
-          exclusive={false}
-          defaultActiveIndex={defaultIndexes}
-        />
-      </>
-    );
-  }
-}
 
 export default class DocumentDetails extends Component {
   constructor(props) {
